move template rendering to pkg
