+++ service/controller/clusterapi/v29/resource/tcnpf/template/render.go
@@ -3,7 +3,7 @@ package template
 import (
 	"github.com/giantswarm/microerror"
 
-	"github.com/giantswarm/aws-operator/service/controller/clusterapi/v29/templates"
+	"github.com/giantswarm/aws-operator/pkg/template"
 )
 
 func Render(v interface{}) (string, error) {
@@ -12,7 +12,7 @@ func Render(v interface{}) (string, error) {
 		TemplateMainRouteTables,
 	}
 
-	s, err := templates.Render(l, v)
+	s, err := template.Render(l, v)
 	if err != nil {
 		return "", microerror.Mask(err)
 	}
